:recycle:  refactor: useLottie hook hooks 폴더로 이동
